fix: unite both code view options under single story file
