+++ src/components/reusable/codeView/blockCodeView/blockCodeView.scss
@@ -1,4 +1,4 @@
-@use '../../../common/scss/global.scss';
+@use '../../../../common/scss/global.scss';
 @use '~@kyndryl-design-system/shidoka-foundation/scss/mixins/typography.scss';
 
 $transition-duration: 0.3s;
