+++ src/components/reusable/codeView/blockCodeView/shidokaDarkSyntaxStyles.scss
@@ -4,11 +4,12 @@
  * @author Hsu Ka An
  */
 
-$default-dark-bg: var(--kd-color-background-ui-strong);
-$default-darker-bg: #272727;
+$default-dark-bg: #272727;
 $default-dark-font-color: #FFFFFF;
 
+
 .shidoka-dark-syntax-theme {
+
     background: $default-dark-bg;
 
     code[class*="language-"],
@@ -22,26 +23,6 @@ $default-dark-font-color: #FFFFFF;
     code[class*="language-"] ::selection {
         background: $default-dark-bg;
     }
-}
-
-.shidoka-darker-syntax-theme {
-    background: $default-darker-bg;
-
-    code[class*="language-"],
-    pre[class*="language-"] {
-        background: $default-darker-bg;
-    }
-
-    pre[class*="language-"]::selection,
-    pre[class*="language-"] ::selection,
-    code[class*="language-"]::selection,
-    code[class*="language-"] ::selection {
-        background: $default-darker-bg;
-    }
-}
-
-.shidoka-dark-syntax-theme,
-.shidoka-darker-syntax-theme {
 
     code[class*="language-"],
     pre[class*="language-"] {
