+++ src/components/reusable/codeView/inlineCodeView/inlineCodeView.scss
@@ -1,7 +1,6 @@
-@use '../../../common/scss/global.scss';
+@use '../../../../common/scss/global.scss';
 
-$default-dark-bg: var(--kd-color-background-ui-strong);
-$default-darker-bg: #272727;
+$default-dark-bg: #272727;
 $default-dark-font-color: #FFFFFF;
 $default-light-bg: #F9F9F9;
 $default-light-font-color: #3D3C3C;
@@ -20,11 +19,6 @@ code {
     font-family: monospace;
     font-size: max(calc(var(--inline-snippet-font-size, 14) * 1px), 14px);
     border-radius: 2px;
-
-    &.shidoka-darker-syntax-theme {
-        background: $default-darker-bg;
-        color: $default-dark-font-color;
-    }
 
     &.shidoka-dark-syntax-theme {
         background: $default-dark-bg;
